requirements.txt
Flask
numpy
pandas
Pillow
scikit-learn
matplotlib
seaborn
joblib
gunicorn